Adds unit tests for find functions
Yay we caught our first bug with a unit test. We were not
excluding functions that start with an `_`. This unit test is
also a little non-standard, in that it tests a few things in one.

I thought this would be easier to manage than creating different
specific files to import… but maybe that might be the better thing
to do.

Note: it's likely that this function will get more complex should
we decide to start decorating functions etc. Right now it's pretty
simple.