Adds unit tests for add_dependency and create_function_graph
I made the decision to add a custom __eq__ (and __ne__) method to the Node
object. This is to make it easier to compare nodes, specifically right
now for unit testing. But I assume this will come in handy later
down the road. Note: if you want to compare whether two nodes
are actually the same thing in memory, then you use `is` to check
if they are the same.

Otherwise the unit tests just construct part of the function
graph in the dummy_functions file - so the graph is hand
rolled for the unit tests -- hopefully this won't be that
burdensome to maintain.