Adds more unit tests for graph module
This adds tests to exercise execute() and get_required_functions().

This I think is the base coverage we should have on the graph module.
What is untested is a few functions in the Node class (no logic), and
the display graph function. I think those are fine to go without coverage
at this stage since I think things could evolve a bit still, and they
don't really have much logic to test.