Flips parameterized decorator and adds requirement to handle doc strings
1. We flip the order of what’s passed in to the parameterized decorator. We did
this because after discussing, we think it’s more natural for people to think of the
column name first, and then to specify the value, then go the other way around.
E.g. more people think about the event name, then necessarily the date that it happened.
2. We add the requirement to provide documentation. This is necessary if we want to
create DRY code and also ensure that people document it. That way when we come
to display/use it all the functions have doc strings.