Adds ability to provide docs to extract_columns decorator
To make use of the documentation available to be stuck into
a Node object, this by default pulls it from the base function. If a
user wishes to override it, then they would provide a tuple of
(“name”, “doc string”) and then that doc string will be used instead
of what’s in the base function.

We do not require documentation for extract columns, since it’s likely
the base function doc string and the name of the column will be
enough documentation for people to understand what is going on.