Adds a @does functionality modifier that allows reuse of generic functions.
The goal of this is to reduce verbosity and improve readability.
We do this by adding a @does annotator that takes in a function
and makes the function its decorating do that functionality instead.
Its currently just a prototype -- it works in limited cases
the replacing function has to only take in **kwargs and function on this.
One could easily imagine it getting more complicated/powerful.