Adds fill_value argument to @extract_columns
There are multiple cases in which we expect to extract columns, but they don't exist.
This includes:
- State-transitions -- an aggregate/groupby with no data won't produce a column
but we still want to output it.
- client count groupbyes -- same deal -- in this case a pivot doesn't exist for some configuration
To make migration to new configurations easier we've added a fill_value argument
that creates a uniform, default series in that case. Note that it also modifies the underlying dataframe in
extract columns. Note this does not solve the full configuration problem of columns that should be in
one configuration and not another -- we'll kick that can down the road a bit.