Pads non-existent columns in kids business line with 0 to match master_data better
The problem is that the groupbys/pivots were not producing data for the kids
business line. It is not clear (for now) whether this is intended or not,
and how we should approach this. Our initial approach is to pad the functions
that are breaking with zeros (so they match master_data closer), but we will
revisit that in the future when we get a better sense. This is a stop-gap for
migration.