Implements a unified configuration API
This enables the following use-cases:

(1) Overriding config parameters with custom json files
    - Allows a user to quickly run a simulation
(2) Overriding config parameters with custom hamilton DAGs/python files
    - Allows a user to quickly inject a scenario/column/parameter to chort-circuit the DAG
(3) Default json files as config
    - Allows us to store config data in code
(4) Default python files as config
    - Allows us to store default config data in hamilton DAGs
    These might be:
        - pieces that grab from S3
        - pieces that grab fields from the json configs (E.G. segment_filters.business_line)
        - pieces that depend on some user-input (E.G. segment_filters)

While including DAGs in the config is complicated, it is a low-enough level primitive that it
enables most things we'll need to do later. If you're confused, the way to think about this is that
all config items are basically sources in the DAG. We've just separated them out so that we can use them
statically rather than dynamically, allowing us to inspect the DAG/inputs.