Adds config as argument to FunctionGraph constructor
While we had previously assumed that a configugration dictionary would only be
used for providing inputs to the function graph, we now realize that the function graph
might need a config to determine edges/connections. This will be particularly useful in:
1) Config-driven models (in which the configuration parameters determine the in-edges)
2) Config-driven columns (in which certain columns might exist for certain business lines)

Although we may not decide to implement (2) ever, (1) is immediately necessary so we are
moving ahead with this change.