Implements + tests model base class
The goal of this is to implement a base class for models and decorators
to enable something like this:

@model(GLM, 'glm_params')
def my_column() -> pd.Series:
    pass

glm_params is the field in the configuration corresponding to the config
parameter that determines how this model is wired. This annotation will
automatically wire up the model to the right inputs in the DAG
and replace the functionality of my_column with a generalized linear model.