Adds implementation of GLM and ongoing cancellation model.
The goal is to enable easily calling models whose parameters are stored
in configuration variables. This implements two of them -- the glm,
and the ongoing_cancellation_model (a subset of glm).
This is partial progress, but it works for the next few commits.
We do not yet support certain model parameter sets (such as
LOG_BAYESIAN and LOG.