Implements columns that predict ongoing cancellations
These are historical columns that use models to predict cancellations
we have not yet received. To do this, we make use of the @model
interface, although this implementation is very similar to the R
code. It may be worth revisiting, as this is a horribly complex
solution, and it is unclear the simplicity of the underlying problem.