Implements @config.when decorator for configuration-driven columns
This allows us to have programmatically-driven column relationships.
To do this, we create a new abstraction, FunctionResolver, that
allows us to resolve a function before modifying it. While this design is not
 perfect, this unblocks the use case we need.
Current limitations:
- No two functions can have the same name (or they'll be overwritten)
- We can't have a default function implementation. E.G. we have to cover the full space
of configuration possibilities through config.when() decorators. To implement this properly
we'll have to have a better set of abstractions and put more thought in.