Adds ops_forecast column
This required fixing how we think about segment filters in the other code that I was
writing. Since we don’t assume that segment filters have region just yet.

Otherwise this seems to match exactly with the columns coming out of the ops_forecast
related data. Only thing that we don’t support yet for this is segmenting by “size”.

Future thought: we need to better handle segment filters and ensure that downstream functions know what
to use/pull out.