Allows nodes to specify their own source
Previously we had nodes either be "user_defined" or not. Now we add
a third choice (NodeSource.PRIOR_RUN) that
enables us to avoid circular references (in the case of a time-series DAG).
While this is a slight change to the original framework, it will save a ton
of code for the time-series implementation.