Adds raw_execute function to hamilton Driver class
There are functions we’ll want to execute that are not conducive to the
shape of a DataFrame. This abstracts the execution out, and provides
a `raw_execute` function, that returns a dictionary of everything that has
been computed. That way things that are not in the shape of a data frame can
be handled more easily.