Adds ability to sanitize config parameters passed in from R
R actually passes in a data frame, where each row is a key
value pair of config. The config is stored in JSON format.

This change inspects the object type, and sanitizes it for
use in python. i.e. we transform the dataframe into a dictionary
and JSON deserialize the values.