Fixes problem of adding multiple handlers
If you’re running code from R you will inevitably run into the case where multiple handlers
are added. This is because we call `log_setup` each time from R. This change
checks the amount of handlers and so if for some reason there exists handlers, we
don’t add any new ones. This might be brittle for other contexts, but it’ll work for now.