Implements dynamic configuration loading
Previously we were evaluating the entire config before executing the DAG.
Thus, items we evaluated were not actually being used. With this change we
evaluate these items dynamically to ensure that we don't have to pay the cost when
we don't need to.

Furthermore, we've narrowed down the space of configuration possibilities. We no longer
allow default json configs, to make it simpler. We can always add back in if we'd like.