Touches up the package settings (#13)
* Touches up the package settings

1. We were not adding the license to the package.
2. This adds a few optional fields that will be useful to have.

* Updates circleci with pre-commit check

This check will ensure we have uniform formatting in the repo.

Otherwise we should depend on circleci images to run the
actual tests.

* Fixing up minor formatting issues

We somehow checked in a few things without it adhering to our
formatting style guide for the repo. This fixes that.

* Bumps version to 1.1.1

So that we can push this package into the wild!