Implements opinionated decorator lifecycle
This adds to the set of decorators that can be applied to a node,
and puts them in a uniform setting. The ordering is provided in
function_base.resolve_nodes -- the algorithm is as follows:

1. If there is a list of function resolvers, apply them one
after the other. Otherwise, apply the default function resolver
which will always return just the function. This determines whether to
proceed -- if any function resolver is none, short circuit and return
an empty list of nodes.

2. If there is a list of node creators, that list must be of length 1
-- this is determined in the node creator class. Apply that to get
the initial node.

3. If there is a list of node expanders, apply them. Otherwise apply the default
nodeexpander This must be a
list of length one. This gives out a list of nodes.

4. If there is a node transformer, apply that. Note that the node transformer
gets applied individually to just the sink nodes in the subdag. It subclasses
"DagTransformer" to do so.

5. Return the final list of nodes.