Builds an expression transform object
This allows us to run nodes like the following:

@transform('a=a+b+c')
def a():
    ...

And now a will be the result of adding a to b to c. Note
a, b, c, currently have to be a node or a config item.

Some caveats:
1. Its unsafe! We use eval now
2. Its just called transform.

Re (2) we should make it so the transform utilizes the same pattern as config:

1. The transform takes in a lambda that takes in a dict
2. transform.to(...) takes in an expression and creates (1) by passing to a superclass