Adds RayGraphAdapter
This code is MVP to get things running on Ray!

One reason why the base.ResultMixin are static
functions on an object, is so that Ray can serialized them.

Otherwise I believe this is an easy way to run hamilton
on a multicore system locally -- as well as connect
to a remote cluster.

The hello world is practically identical to the dask one,
apart from setting up the GraphAdapter and ray config.