Refactors model example to show how to be generic
This example is interesting because it shows how one might
build a "bank" of hamilton functions that do some generic
modeling -- while keeping it generic so that adding new contexts/
running it with new models, results in a small amount of work.

The key things to get this to work are:

 - different python modules to load data. They have to output what's required to link
   with the my_train_evaluate_logic functions.
 - config & @config.when to add the correct model function to the dataflow.

So if you want to switch between model types: easy -- change config.
So if you want to switch fitting models on different data: easy -- change the data loading module.