Adds numpy example from a numpy tutorial on doing AQI
See https://github.com/numpy/numpy-tutorials/tree/main/content
for where this is pulled from.

The names of the functions could be better I think. Otherwise this
doesn't seem like a bad pipeline. Versus doing this analysis in a notebook,
the strength of Hamilton here is in forcing concise definitions and language
around steps in the analysis -- and then magically the analysis is pretty
reusable / very easy to augment.

E.g. add some @config.when or split things into python modules to be swapped out.