Refactors type checking logic to help fix #91 and implement #39
We were not doing a great job ensuring that the issubclass check would not break.
So this changes that, and adds more use of the typing_inspect module. We'll have
to test this for compatibility to ensure things work with 3.6.

So I specifically created a dedicated function to encapsulate this type checking logic.

To handle generics, I did a very basic approximation that should work for most use cases.
E.g. list vs typing.List, or typing.Dict vs typing.Mapping.

In terms of precedence, the param_type is the input parameter type annotated on the function,
and the requested_param_type is coming from the output of an upstream function. Thus, the
param_type should be the super type if there is one, not the other way around.