Enables full support for optional dependencies.
This was handled but not fully accounted for. The changes we made were:

1. Add compile-time checking of optionals (E.G. optionals don't have to exist)
2. Add runtime checking of optionals

To do (2) we had to change traversal of the function graph to account for
runtime inputs. This is because, in the driver, we ask for all functions
that are dependencies and then run them. We know what inputs we have,
so we can ask for all dependencies that we're going to need for a specific
execution. Thus we never ask for dependencies that don't need to exist

We should probably not always return the intermediate values, but we can
deal with that the next time we dig into materialization.