Adds some basic graph query functions to driver
Thinking of how to use tags, or the workflow to use them, I'm envisioning

1. query for all nodes.
2. filter by tags.
3. use those nodes then for other purposes.

Ideally we could instead provide an API that you use tags to
query by instead, but I don't have a good feel for what that
should be right now.

Anyway, I'm thinking it'd be useful to ask questions,
given a set of function names, what is downstream or
upstream of them.

I only added visualization of the downstream because,
you can visualize what's upstream by using the
visualize_execution function.