Changes function_names to node_names
Okay, going back on this one, Elijah convinced me... But I think we have a mess as it is with naming
things in the driver. The most neutral is "node" since we talk about DAGs all the time, and using
function_name would be introducing yet another parameter name concept.