Adds the ability to pass functions not defined in modules - fixes #134
We went with the decision to create an object to house functions for three reasons:

1. It meant we did not need to structurally change the underlying code, other than checking
for the object instance type, and amending some type annotations.
2. This adds some friction to force people to at least acknowledge they need to curate the functions.
The name `TemporaryFunctionModule` should indicate that this should not be a production thing.
3. This opens the door for us to change our mind in the future about how we do this.

The `my_notebook_script.py` shows what one would do in a notebook context,
i.e. shows how you can inline define functions that interact with a module that constructs a DAG.

Assumption we did not break:

 - you cannot define the same function twice -- so you cannot use this as a way to override a  previously
defined function.

Adding this for context, if you want to overwrite a function implementation you instead have the following paths:

1. Choose the subset of functions you want and put them in the TemporaryFunctionModule, i.e. import those functions.
2. (a)Run the upstream part of the DAG to create the inputs required for the new function.
(b) Then compute the result with the new function. (c) Then run the rest of the DAG using the `override=` option of execute.