Adds pandera integration for data quality
This is a pretty simple approach but I think it works nicely.

We use the schema= decorator to specify data quality checks
that validate a pandera schema. Note that this will only be registered
if pandera is installed (which is an option for setup.py).

The other option we were thinking of is to compile the default checks to
pandera, but I think that's a little overkill, and can be done later.

Furthermore, this proves out the abstraction for data validation quite
nicely -- this was straightforward to implement.