Proof of concept that we can easily add a profile step.
What this does...

1. Adds a new profiler argument in data validation decorator
2. Adds scaffolding for a whylogs class
3. Messes with DAG validation to allow for a profiler. Note that if we
   don't have a profiler, the DAG just runs validators on the original
data. If we do, it runs validators on that. The validators then have to
accept the type the profiler outputs, rather than the original data
type.

What this is missing

1. Testing -- its just a POC with pieces left out. Run at your own risk
   :)
2. Configuration-wiring -- hopefully that'll come soon.
3. Tag-exposure -- that will be useful for giving metadata to the
   decorator. The data is there, but its not available yet.

Note there are other approaches -- I think this is nice as its flexible
yet opinionated, and we can always go down to the lower API if we have
something more complex.