Refactors validator `name()` function and renames validator
Sorry, merging two commits here.

(1) The `name()` was pretty much just the `argument` + `_validator`.
So I just encoded that and updated the names of variables, and classes
to match this format of doing things.
Thus (2) is rolled into this. Because we should be making sure that
arguments, names, class names, are following some semblance of
structure.