Refactors BaseDataValidationDecorator and act/warn
Prior behavior stopped at the first failure. We don't want
that to happen. Instead we want to run through all the
checks and log them appropriately, this change does that.
So I had to changed "act" to accommodate this.

Since `act` itself was only used in a single place, I just
moved the `if` statement into the BaseDataValidationDecorator.

That said, the class structure here feels a little odd -- might
be easy to introduce a circular dependency at some point accidentally.
But yeah we need a better mechanism for storing results
for people to access.