Fixes a bug in which deprecation was called on function definition
It should be called on function call. This, however, adds complexity,
as we need to be able to decorate both classes (that define __call__)
and functions (that happen to have __call__). There is some complexity
that python doesn't handle well -- a third party library like wrapt
might help but I wanted to keep it simple.