Remove the typing constraints for the replacing function in @does
 This enables more complex polymorphically applicable functions without
the pain of implementing sophisticated type-checking.

Furthermore, I think this was overreach by the Hamilton framework.
Hamilton cares about relationships between nodes, and not the
implementation of nodes. As `@does` creates a node, this does not impact
the relationship between nodes, meaning that this type of type-checking
is out of scope.