Refactors decorators to live in a separate subpackage
This makes it backwards compatible -- we previously had two files:

1. function_modifiers.py
2. function_modfiers_base.py

We now create a module `function_modifiers.py` which enables us to
separate out decorators, as the file is getting unwieldy and ugly.

That said, we need to maintain backwards compatibility on import.
To do this, we:

1. Create a function_modifiers subpackage
2. Put everything that was in function_modifiers.py in __init__.py in
   that package (so imports work)
3. Move function_modifiers_base.py over there
4. Create a "dummy" function_modifiers_base.py that just imports
   everything from the actual one.

Will need to test this out on sample workflows prior to publishing.