Adds unit test for pandas index with DF creation
TIL: you can create a dataframe and pass in an index object
and it'll happily use it as a column.

So this test should exist for dataframe creation since it's a valid case.

But for the index type checking, I'm adding it here even though
it does not have an explicit index. Therefore, one could make the argument it
doesn't qualify here. But, I'd rather push people to be explicit
in their code, e.g. if they want to be strict on indexes, then they should
make the index a series, rather than passing an Index object.