Example showing what we would need to modify to enable user question
If the user wants to reuse hamilton functions without hamilton, this
commit shows one way to do it -- and what would be required on Hamilton's
side to support it.

Basically we'd need a convention. Not sure it's a good idea. Probably
easier to instead have boilerplate code people use to try/except the
import if hamilton does not exist and have the decorators be
identity functions...