Fixes @does interaction with @extract_columns #244
Cleaner fix for `@does`. The function is only useful for the node generator
 -- not for the rest of DAG creation. This ensures that we don't use it
to handle signatures, rather, we just use the node's return types.

This also adds unit tests, and in the process opened #249 since we need
better clarity and test coverage on decorator interactions.

---- Squashed commits ----
* Fixes does function wrapping #244

In issue #244 types were not being propagated of
the function being wrapped.

This changes that by adding the functools.wraps
decorator.

Adds a unit test to test for two decorators playing nice
together and verified that it's broken before and fixed
afterwards.

* Removes print statement

* Cleaner fix for `@does`. The function is only useful for the

node generator -- not for the rest. This ensures that we don't use it
to handle signatures, rather, we just use the node's return types.

* Adds unit test to cover extract_fields with does

Adds a test to cover Elijah's changes.

In the process I created https://github.com/stitchfix/hamilton/issues/249.
Since I think we need to add more unit tests around decorator interactions.

Co-authored-by: Stefan Krawczyk <stefank@cs.stanford.edu>