Adjusts index type check warnings (#246)
* Adjusts index type check warnings

Following #243 we need to adjust the warning
to make sense. Since we are not in danger of things
breaking imminently. So the warning is just there for
the user now to indicate how Pandas might be resolving
indexes to create the dataframe.

* Fixes bug with showing count of no index outputs

This message was just wrong with the count displayed.
So fixing it to reflect how many no index outputs there
actually are.