Fixes unit tests to be more robust
Async test was breaking because of the extra
telemetry invocation. Adding mock context manager
ensures that it is disabled for that one invocation.

Otherwise permanently fixing the sanitize error
issue by regex replacing the line number, since that
will be changing depending on the number of
lines in telemetry.py and we don't want people
to have to update this test because of that.