Adds polars example
Polars is a competing dataframe library. This is a minimal
example based off the pandas hello world that shows
it's pretty easy to right polars code with Hamilton too.

Note: we don't use the `select` syntax here. I don't know
whether this is a best practice or not. But hopefully someone
from the polars community could help set us straight here.