Prototypes some code
To show how we might not require users to pass in
the dataframe and series types. Instead we have
some functions that register themselves,
in addition to using singledispatch which will
at runtime choose the right type.

This is conceptual code -- we'd need to clean things
up and add support for dask dataframes, etc.