Makes the quokka hello world run with Hamilton API
This is highly experimental. Just getting a feel for:

1. the hamilton API one would need to write.
2. what we'd need to adjust within Hamilton to make
things work as expected.

There is a lot of hackyness. But basically it seems that
with a graph adapter, assuming correctly node traversal,
then at least for the case of a single datastream object
we can correctly derive things.

Edge cases to handle:

- multiple input datastreams
- joins -- what's the syntax for that?
- schema validation -- we use that to understand what should
or shouldn't be used.

API thoughts:

 - groupby doesn't seem terrible, but really the question is: where
do you house the group by logic? in the adapter? or in the function?
 - need to figure out some join syntax