Adds m5 forecasting kaggle example (#280)
I pretty much verbatim copied the code from [this notebook](https://www.kaggle.com/code/ratan123/m5-forecasting-lightgbm-with-timeseries-splits).
It does have a few odd quirks in it -- so if I understood what was going on I removed it, else I left it in. E.g. `test_2` is not used... It is however known to be a good example of a beginner notebook that shows basic approaches to
modeling the problem.

This is an example that shows:

1. loading data
2. creating a base set.
3. adding more features.
4. creating a training set.
5. fitting model & predicting with it.
6. saving the result.

There are a few other ways one could write some of this code.
But as a first pass this seems fine. E.g. we could split up the model
training function a bit more. Or we could use the parameterize decorator.

# Tweaks I made

1. added doc strings to help people get the feel for that.
2. refactored some functions to make it appear more "production like", showing
better reuse/smaller functions, and better naming. E.g. the fit and predict functions
 should be themselves nodes in the DAG, that a function can itself call out to.

# Notes

No unit tests have been added, or data quality checks -- however for real production
purposes we would expect one to add them.