Adds ability to pass in a callable to get the function name from for … (#283)
* Adds ability to pass in a callable to get the function name from for execute

I found myself wanting to click directly to outputs being created to inspect them.
So to short cut that, directly referencing them is a way I can do that and have
my IDE jump there when navigating code.

The downside is that this solution doesn't handle functions annotated with decorators
that create nodes, e.g. @parameterize*, or @config* . I think that's okay, because
it's probably clearer to have a "string" name here to indicate that something is
changing/being produced at DAG creation time, and thus a reference to an
exact function is likely a misnomer...

Adds unit test for this.

* Adds check for functions to be in modules used in driver

So that way you can't use a reference to a function to a module
that is not used. This should stop cases of copy paste erroring/changing
when people change something in the referenced function.

* Updates hello world script to show function references

I think this might be the better default going forward? Let's try it
out and see.

I also replaced the module import, by directly importing the function
module and commenting that out. I had received feedback that it looked
complicated, but I just wanted to show you can script the import of modules,
so the driver script code could be argument driven...