Changes visualize driver functions to return graphviz object (#286)
* Changes visualize driver functions to return graphviz object

This change is backwards compatible. It returns and wires
through returning the graphviz object created. That way in a notebook,
it can render the DAG for you.

It also makes saving to a file optional -- you just need to
pass in None explicitly. Adds test for it.

The hello world example is also updated with a real notebook
and shows Thierry Jean's approach of using the autoreload
extensions.

* Adds one more assert to test