Decorators now declare the configuration parameters they need
This allows for inspection/visibility, and will be essential in
checkpointing/storage options. We want decorators to declare these
upfront, rather than relying on the internals of the decorators to
break, enabling clearer error messages for more usable
configuration-driven pipelines.

Furthermore, we also allow for some "optional" configuration parameters.
These enable us to provide a default. The use-case here is config.when
-- which currently defaults to None. Allowing this from the start is not
ideal, but it's what we did and users rely on it.

There *is* an out, however. If someone is using the base `@config`,
they were able to declare a lambda function, which yields no insight
into the parameters required by the function. To bypass this, we pass
None out, and the framework knows not to filter the required parameters.