Adds code to make concatenating dataframes columnwise in the result builder default (#321)
Why this change? because if you are doing indicator variable stuff,
maybe you don't want to be explicit about them, instead you want
to concatenate the columns created without knowing about them. Seems
like a reasonable pattern on first glance.

This code works because it delegates to pandas on how
to construct everything -- all it does is that it flattens
the dataframe, and then ensures there's no duplicate column
definitions. I added tests for this for this case. One could namespace
the columns and make it behave pd.merge like, but I decided that we
should punt on that until there's a good use case for it.

For the future, we probably want some flexibility with respect
to "joins" and how things are stitched together since you could
envisage some use cases where people want to inner join between
things --- but until there's a common pattern/need, I think
we should just keep thinking about what should be first class,
and what we get people to do themselves.