Adds extensions and path updates for conf.py
Added extensions are for creating reference docs from our python
files easily.

Otherwise it adds the hamilton repo to the path when sphinx consumes
conf.py. This will mean that the hamilton code is now accessible to be
referenced for when we want to create documentation from those modules.

Note: sphinx will need to have all the right python dependencies in order
to be able to parse that python file.