Adds PySparkUDFGraphAdapter to h_spark
This is a large commit from a couple of squashed ones, original commits below.

The introduces a new experimental graph adapter - PySparkUDFGraphAdapter.

The point of this adapter is to allow someone to build and manage a DAG of map functions that would result in columns
on a pyspark dataframe.

def spend_per_signup(spend: pd.Series, signups: pd.Series) -> column[pd.Series, float]:
    """The cost per signup in relation to spend."""
    return spend / signups

def spend_per_signup(spend: float, signups: float) -> float:
    """The cost per signup in relation to spend."""
    return spend / signups

Limitations:
 - Only works for "map" operations.
 - Does not implement the full spectrum of pandas_udfs available. Not all are map based.
 - There potentially are performance impacts to using with_column for a large DAG. We should look to use expressions, and then do one final select statement in the result_builder.

——
Pulls out static functions from PySparkUDFGraphAdapter

They don't need to be in the graph adapter itself. Pulling them
out makes them easier to unit test, as well as simplifies the
class itself.

Adds unit tests for the basic functionality of each static
function, as well as an integration test for the overall
adapter. (+5 squashed commits)
Squashed commits:
[857b91c] Adds to PysparkUDFGraphAdapter docs and functions

1. h_spark, so that they're more understandable and chunkable.
This should make it simpler to unit test them.
2. sphinx docs, so that this class shows up there for reference.
[ff60fb1] Adds future work section to pyspark udf readme

So that it's clear we have more work to do here.
[7558cfa] Fixes UDF driver reuse bug

We need to depend on type information from the node, not the actual callable itself, since we
will mutate the annotations on it for running spark. This just impacts behavior of rerunning things.
[a08c392] Adds PySparkUDFGraphAdapter

TODOs are to fill in docs and add tests.

Otherwise this is functionally ready to go!

Regular UDFs are pretty straightforward.
Pandas UDFs involve a bunch more code, and we only support one type of them.
So the bulk of the complexity is in supporting them and massaging
the function because pyspark expects no extra annotations etc.
[b76bbd5] Adds WIP code to run pyspark UDFs through Hamilton

The current behavior assumes:

1. a single DF being operated on.
2. the execute function says what should additionally be added to the current dataframe.
3. you need to do joins, filters, aggregations, outside of UDFs.
4. the pandas_udf stuff works, but assumes no ability to pass in scalar values to the function.
So you have to instead have them be in the columns.

TODOs:
 - move files
 - more testing
 - unit testing
 - adding usage documentation and limitations.