Adds support for from __future__ import annotations
Luckily the code we have is fairly clean -- this doesn't have a big
blast radius. In reality, it should be one spot we make the changes, but
we'll have to refactor this later.

With __future__ import annotations, python now allows you to lazily
evaluate annotations. In <3.10, there is no way to get an
inspect.signature with fully evaluated annotations, so we have to use
typing.get_type_hints. In 3.10, there's a parameter to force-evaluate
signatures, so we can switch to use that when we deprecate 3.9 in quite
a while.

This is tested in the smoke-screen module -- testing is tricky due to
complexity of "future statements". We have to hack up imports, so I
chose only to do that in one place.