Adds basic scenario 1 for feature engineering
Assumptions:
 - the API request can provide the same raw data that training provides.
 - if you have aggregation features, you need to store the training result
for them, and provide them to the online side.

This example shows how one might use Hamilton to compute
features in an offline and online fashion. The assumption here
is that the request passed into the API has all the raw data
required to compute features.

This example also shows how one might "override" some values
that are required for computing features, in this example they
are `age_mean` and `age_std_dev`. This can be required when you
computing aggregation features does not make sense at
inference time.