Fixes #93
This is a less than ideal solution, but basically if the function being wrapped is a coroutine,
then make the wrapper async as well. This duplicates a lot of code. I'm sure there's
a more succinct way to do this, but because I'm time pressed I'm doing the more
verbose solution.

Note: this doesn't fix it for all decorators, just the extract* ones.

Adds async tests to double check and ensure that things work as expected.