Adds dynamic node
See #109

Note this makes the following decisions:
1. Chooses to use the generic `@dynamic` decorator
2. Buries it under a power-user mode
3. Uses the parameter names of the labmda/function to draw from the
   config

We also Deprecate dynamic_transform

This was an old way of doing the same thing. This was never accessible
to users, as it reuqired implementing a subclass and konwing some of
hamilton's internals. delay_resolution solves the same problem, and
gives users access to any of the current decorators.