Ensures power-user mode does not get filtered and gets passed to decorators
Every decorator should have access to this, just in case, so we don't
have to worry about passing it through. This changes the resolve_config
function in function_modifiers.base to pass it through globally. Note
that this also *removes* the default setting of this in the @resolve(...)
decorator as it is no longer necessary.

Furthermore, this cleans up some issues in which the optional configs
provided by a decorator was a list instead of a dictionary.