Fixes a bug in which nested subdags were not passed required configuration parameters
We're aggressive in pruning parameters we pass to decorators, so they
don't use everything. SubDAGs didn't declare them, as they only know the
configurations required once they parse the function that they decorate,
as well as import the subdags they use.

Rather than the more correct, but more complicated fix, this utilizes
the "out" -- allowing required_config to return None so everything gets
passed in. We need to think about the best way to do this, but for now
this is little harm and we can walk it back -- the contract is solid.