Exposes more information for check_output decorator errors
If the user does something that involves the mistyping `target_` or
tries to use a key word argument that doesn't map to a validator for the given
output type, then the error will be cryptic. This adds context that it's
for the check_output decorator and which function it's for.