Simplifies logic for parameterized decorator
This has been a long time coming. Previously, we were looping through
all the replacements and modifying the kwargs. Now, we just loop through
all the expected inputs, and create their results from the kwargs or the
literal values. In the case of "grouped", we do the same -- we just
repeat some of the previous stuff. There's room for more simplification
(breaking this out into a function), which could reduce the complexity
of the code, but for now its OK.