Fixes regression introduced by enforcing the config on the `@config` base decorator.
This decorator is almost never used -- its specifically built to have
arbitrary configuration elements passed through to a resolver function.
That said, some users found it, and it turns out it was hitting a
codepath that results in a null pointer. This changes the node
resolution code to bypass the prior checks we had to ensure that
all configuration elements were passed in, and handles the case in which
the passed configuration elements are null (as is with `@config` and no
other decorator)