Fixes issue with @inject and docstring
Beforehand, we broke if there was a docstring in an inject decorated
function. This was due to having a placeholder function name, as it
delegated to `@parameterize`. The best fix would be to not have it
delegate (this is a bit of a hack), but for now we make it not break.