Fixes dangling node in subdag issue
If there was a subdag that carried a dangling node (E.G. one that was
not depended on by something within the subdag), and it was used by
'extract_fields', this would break. This occured because the
extract_fields decorator did not know which node to extract fields from,
as it looked for "final" nodes within the whole set of nodes produced by
the function, including those within the subdag.

This fix allow `@subdag` to modify it so we don't have this issue
anymore.

Currently, the way that the subdag indicates that its nodes are
"non-final" (e.g. they can't be the target of extract_columns) is
through a decorator. This can always be changed, but its not a bad way
to do it. Furthermore, this will happen *after* the resolution of any
subdags, ensuring that nesting subdags still works, as does any
decorators within subdags.