Exposes package version as __version__
This seems to be the most defactor standard way of exposing package
version information.I decided to keep it as a tuple, since if you want to compare/pull things
from it, it's already split into numeric numbers. There doesn't
seem to be consensus here on whether it should be a string or tuple. But
my guess is if you're programmatically pulling it, you'll want
parse the string anyway, in which case, why translate to a string
in the first place.

Thanks to @sT0v for the push here.