Adds the original function pointer to node/variable objects
Before we had no way of knowing which nodes were generated by which
functions. This adds the capability. Each variable now points to the
functions that generated it.

Note this is a tuple of functions as multiple functions could have
generated it (in the case of a subdag). The order of these will always
have the lowest level function first, and will likely correspond to
namespaces, in reverse (although this contract isn't ironed in yet).