Fixes #161 - adds node name to hamilton dask DAG (#163)
So dask was taking the names of the wrapped function, rather than
the name of the actual node. So to keep things unique, we use the
recommended tokenize function to hash the inputs to create
a unique string. We use it for both name and dask_key_name because
our functions only depend on the input and aren't called multiple times.
So we should be good to reuse the same string value in both places.

---------

Co-authored-by: Stefan Krawczyk <stefan@dagworks.io>