Removes need to separate user_nodes in visualization
Splitting into user nodes is now obsolete with the node_modifiers
dictionary. So this refactors the code internally to remove the `user_nodes`
argument, and instead pass in all nodes with the appropriate node modifiers.

I seemed to also fix a bug that's been there in the dot file metadata -- in that
the test was adding two nodes for the same function.