Adds node name anonymization for temporary generated nodes
Previously we had hardcoded rules that could occasionally clash. This
anonymizes it every time. While this isn't the prettiest way to do it,
its better than implementing state. Next up, we'll be building out
tooling to mark these with tags as related to a cluster of nodes, and
use that to group the DAG.