Renames graph get_impacted_nodes to get_downstream_nodes
We're doing this for consistency, since this complements `get_upstream_nodes`.

This is fine to change, since the `graph.py` module is seen an internal API
and not a publicly facing one.  But just in case someone is somehow using it, I'm leaving it in, and
delegating to the new function with a warning message.