Extends the cache to work more generally
This packs it all into h_cache.py for now.

It uses singledispatch because that's an easy way
to register functions for different object types.

In terms of deserialization, we hack things, by including
an empty expected type to get singldispatch
to read from the right function, since singledispatch
uses the first argument to resolve which function, and it
only works in instantiated types, and not on a Type object.

Longer term we should merge this with the datasavers --
i.e. the data savers should in general have an easy way to
access SERDE functions, but without the boilerplate overhead...

CAVEATS to make this truly generic is:

1. How to register these across different object types. Should these
be defined in the extension classes?
2. KWARGS: we need to handle how to pass these through nicely.
For the adapter, do we use tags? or allow some other means to inject
kwargs? Ditto for getting this working with the data savers.