Caching Graph Adapter (#195)
This adapter is "experimental", and being shipped to get feedback from the community on it.
It allows one to cache a node (invariant to inputs) and extend how it is cached and to what.

Thanks to @elshize for the idea and design. It came about during the need
to iterate on a Hamilton dataflow (i.e. pipeline). The design isn't fancy,
and we can always extend it (e.g. to be dependent based on inputs). 
From an 80/20 perspective, this probably hits the mark and we're excited to 
push this out.

See docs & code comments for more information on how to use this, and caveats.

Squashed commits ---- 
* Caching adapter

* Reset computed node state when collecting results

* Import module instead of functions directly

* Change doc format

* Document `execute_node`

* Clarify docstring

* Extends the cache to work more generally

This packs it all into h_cache.py for now.

It uses singledispatch because that's an easy way
to register functions for different object types.

In terms of deserialization, we hack things, by including
an empty expected type to get singldispatch
to read from the right function, since singledispatch
uses the first argument to resolve which function, and it
only works in instantiated types, and not on a Type object.

Longer term we should merge this with the datasavers --
i.e. the data savers should in general have an easy way to
access SERDE functions, but without the boilerplate overhead...

CAVEATS to make this truly generic is:

1. How to register these across different object types. Should these
be defined in the extension classes?
2. KWARGS: we need to handle how to pass these through nicely.
For the adapter, do we use tags? or allow some other means to inject
kwargs? Ditto for getting this working with the data savers.

* Adds pyarrow to test dependencies

So that we can test the caching serialization.

* Fixing h_cache annotations for 3.7 & 3.8

Primitive types aren't subscriptable until 3.9

* Adds more documentation to explain h_cache set up

* Correct writers/readers type hints

* Add examples in docs

* Logging messages

* Reformat

* Add reference entry to CachingGraphAdapter

* Add example project

* Update examples/caching_nodes/README.md

Co-authored-by: Stefan Krawczyk <stefan@dagworks.io>

* Fix caching adapter example

* Finish documentation

* Adjust requirements.txt for caching example

So that people can run it without problems. By default
pyarrow is not installed, so the example breaks.

Also setting sf-hamilton minimum based on future release.

---------

Co-authored-by: Michal Siedlaczek <michal.siedlaczek@ibm.com>
Co-authored-by: Stefan Krawczyk <stefan@dagworks.io>